pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "estate-scraper"
version = "0.1.0"
description = "Estate sale scraper with AI-powered valuation"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.40",
    "anthropic>=0.40",
    "pydantic>=2.0",
    "typer>=0.9",
    "rich>=13.0",
    "httpx>=0.25",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "tenacity>=8.2",
    "pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
estate-scraper = "estate_scraper.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
